refactor(#163): extract analysis.py — geometry/feature analysis (P4)
Move _analyse and its private helpers (_import_step, dedup_diams, _is_rotational,
analyse_face_levels) out of make_drawing.py into analysis.py. _analyse imports the
part, runs feature detection, classifies rotational-vs-prismatic, and chooses the
sheet + view zones (via sheet.choose_scale/_layout_geometry) — returning the
Analysis namespace the pipeline reads.

analysis.py sits above sheet, below make_drawing: it imports _core, sheet,
build123d_drafting features, draftwright.features (find_slots), build123d/OCP — never
make_drawing/annotate, so no cycle. make_drawing re-imports _analyse (build flow);
the public analyse_face_levels/dedup_diams and choose_scale move with __init__ to
their real homes (analysis / sheet); _MIN_VIEW_MM relocates to _core; the
rotational tolerances (_OD_*/_SQUARENESS_TOL) move into analysis.py. The lint
scoring constants (_SCORE_*/_GEOMETRY_AWARE_CODES) stay in make_drawing with
lint_summary.

Tests now import the moved internals from their canonical homes. Pure move: golden
gate passes against the committed snapshots with no regeneration; full
make_drawing/golden/e2e suites pass (361); ruff check + format + mypy clean.
Adversarially checked: no upward import, no stale defs, rotational/prismatic parts
classify correctly with scale+zones set, lint_summary scores, FeatureInfo's
@dataclass intact.

Closes #163.
